feat: Separate ReviewBox
