+++ src/components/DetailPage/Reviews/ReviewsLayout.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ReviewBox from './ReviewBox';
+import ReviewBox from '../Review';
 import { Review } from '../../../lib/interface';
 import styles from './ReviewsLayout.module.scss';
 
@@ -7,7 +7,6 @@ interface ReviewsLayoutProps {
   count: number;
   reviews: Review[] | null;
   maxPageIdx: number;
-  accessToken: string | null;
   onPageChange: (idx: number) => void;
 }
 
@@ -15,7 +14,6 @@ export default function ReviewsLayout({
   count,
   reviews,
   maxPageIdx,
-  accessToken,
   onPageChange,
 }: ReviewsLayoutProps) {
   return (
@@ -25,7 +23,7 @@ export default function ReviewsLayout({
         <ul className={styles.review_list_wrapper}>
           {reviews?.map((review, idx) => (
             <li key={idx}>
-              <ReviewBox review={review} accessToken={accessToken} />
+              <ReviewBox review={review} />
             </li>
           ))}
         </ul>
